Dev build 176-178 + Rework structure
This commit adds new dev builds: 176-178, which offers OP Prisons S6 compatibility - Complete rework with enchantments. Bug fixes that were originated from S5, and improvements that are being made.
We're also depreciating Classic OP Factions, and making OP Prisons S6. So all the changes relating to classic OP Factions are now gone for the time being. We may add this in the future. But for now, we're working on one server at a time.